(1) float arrays changed to double arrays to mantain consistency within the project
(2) buffer - implementation started
(3) baseline correction implemented
(4) minor refactoring

+++ src/cz/zcu/kiv/eeg/gtn/data/providers/bva/OnLineDataProvider.java
@@ -40,7 +40,7 @@ public class OnLineDataProvider extends AbstractDataProvider {
             Object o = dtk.retrieveDataBlock();
             if (o instanceof RDA_MessageData) {
                 RDA_MessageData rda = (RDA_MessageData) o;
-                float[][] data = new float[channelCnt][(int) rda.getnPoints()];
+                double[][] data = new double[channelCnt][(int) rda.getnPoints()];
                 EEGMarker[] markers = new EEGMarker[(int) rda.getnMarkers()];
 
                 float[] rdaDta = rda.getfData();
